Refactor of the Alert component to make it an error-only alert component.
